Renamed files "Dispatcher". Fixed "dispatch" method. Registered an implementation for an interface.
