+++ bootstrap/services.php
@@ -2,8 +2,10 @@
 
 declare(strict_types=1);
 
+use App\Providers\DispatcherProvider;
 use App\Providers\TwigProvider;
 
 return [
     TwigProvider::class,
+    DispatcherProvider::class,
 ];
